edit yahweh to yahuah
also note: on the repository description, when attempting to change the tags "scripture, yahweh, messiah", to "scripture, yahuah, messiah", github forces the order to be " scripture, messiah, yahuah". perhaps this is because the tag "messiah" was already in place before the tag "yahuah" was added. i tried removing the "messiah" tag to try to force it to order it how i intended, but github's behavior is apparently uncontrollable. i consider that to be a bug with github itself; that the order of tags seems like a user-defined thing, and may be in most cases, but in some cases github does not behave as expected, and instead re-arranges the order.